Fixed reducer usage of rErrorMessageOpen
Note that technically this makes a couple of reducers slightly impure,
due to the uuid of the message.
This is a hack which should be refactored in the future, by generating
the uuid before the reducer.

A uuid could possibly be generated with every action so that it is
always available, but this would require a wider refactor.